Function ostn15_phf::ostn15_lookup

pub fn ostn15_lookup(key: &i32) -> Option<(f64, f64, f64)>

Return a 3-tuple of Easting, Northing, and height adjustments, for use in transforming ETRS89 Eastings and Northings to OSGB36 Eastings, Northings.

Examples

use ostn15_phf::ostn15_lookup;
 
// Caister Tower Eastings and Northings: 651307.003, 313255.686
let e_grid = (651307.003 / 1000.) as i32;
let n_grid = (313255.686 / 1000.) as i32;
let key = e_grid + (n_grid * 701) +1;
// key should be 220065
let result = ostn15_lookup(&key).unwrap();
// result should be (102.787, -78.242, 44.236)
assert_eq!(result, (102.787, -78.242, 44.236));
// remember that the actual adjustment for a coordinate is a bilinear transform, using a square